Make DownloadSelector tasks static and extract them to separate files
